+++ CHANGES.md
@@ -1,5 +1,6 @@
 [unreleased]
 * ensure WP 6.0 compatibility with `move_dir()`
+* use JSON in plugin root for non-dot org dependencies _acceptable_ for dot org 🤞
 * update test plugins
 
 #### 1.8.0 / 2023-04-07